[naga xtask] Move validation subcommands into their own module.
